+++ system/scripts/dice/dietype/ability-die.js
@@ -5,7 +5,7 @@ import { L5rBaseDie } from "./l5r-base-die.js";
  */
 export class AbilityDie extends L5rBaseDie {
     /** @override */
-    static DENOMINATION = "ds";
+    static DENOMINATION = "s";
 
     static FACES = {
         1: { success: 0, explosive: 0, opportunity: 0, strife: 0, image: "skill_blank" },