+++ system/scripts/items/technique-sheet.js
@@ -16,19 +16,15 @@ export class TechniqueSheetL5r5e extends ItemSheetL5r5e {
     }
 
     /** @override */
-    async getData() {
-        const sheetData = await super.getData();
+    async getData(options = {}) {
+        const sheetData = await super.getData(options);
 
-        // Add "school ability", "mastery ability" and customs if active
-        [
-            ...CONFIG.l5r5e.techniques_school,
-            ...(game.settings.get("l5r5e", "techniques-customs") ? CONFIG.l5r5e.techniques_custom : []),
-        ].forEach((e) => {
-            sheetData.data.techniquesList.push({
-                id: e,
-                label: game.i18n.localize(`l5r5e.techniques.${e}`),
-            });
-        });
+        // List all available techniques type
+        const types = ["core", "school", "title"];
+        if (game.settings.get("l5r5e", "techniques-customs")) {
+            types.push("custom");
+        }
+        sheetData.data.techniquesList = game.l5r5e.HelpersL5r5e.getTechniquesList({ types });
 
         return sheetData;
     }